backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
plaid-python>=11.0.0
python-dotenv>=1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
requests>=2.31.0
python-jose==3.3.0
passlib==1.7.4
